Changed the way the Zookeeper locking works.
Locks are no longer remembered and they are released using a token. It should prevent the situation when the releasing for some reason fails and the client still believes that it holds the lock (and it's never removed). Now every lock acquire call actually contacts Zookeeper (before if the lock was not forgotten, the client believed it has a lock and could work without obtaining one, i. e. even in situations like when the Zookeeper was gone).